[build-tools] move credential catalog and error into PosthogUtils
