+++ packages/build-tools/src/steps/utils/PosthogUtils.ts
@@ -5,6 +5,27 @@ import { bunyan } from '@expo/logger';
 import { Response } from 'node-fetch';
 
 export namespace PosthogUtils {
+  export interface Credential {
+    label: string;
+    envVar: string;
+    input: string;
+  }
+
+  export const API_CREDENTIALS = {
+    apiKey: { label: 'personal API key', envVar: 'POSTHOG_CLI_API_KEY', input: 'api_key' },
+    projectId: { label: 'project id', envVar: 'POSTHOG_CLI_PROJECT_ID', input: 'project_id' },
+  } satisfies Record<string, Credential>;
+
+  export function missingCredentialsError(missing: Credential[]): UserError {
+    const labels = missing.map(credential => credential.label).join(', ');
+    const envVars = missing.map(credential => credential.envVar).join(', ');
+    const inputs = missing.map(credential => credential.input).join(', ');
+    return new UserError(
+      'EAS_POSTHOG_MISSING_CREDENTIALS',
+      `Missing PostHog credentials: ${labels}. Set the environment variables (${envVars}) or step inputs (${inputs}) on EAS, or re-run "eas integrations:posthog:connect" with error tracking enabled.`
+    );
+  }
+
   export const DEFAULT_POLL_TIMEOUT_SECONDS = 600;
   export const DEFAULT_POLL_INTERVAL_SECONDS = 30;
